requirements.txt
requests==2.18.4
python_telegram_bot==11.1.0
